refactor: move extension config out of api
