+++ runtime/devnet/src/config/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod api;
 pub mod assets;
 mod contracts;
+pub(crate) mod extension;
 mod proxy;
 // Public due to integration tests crate.
 pub mod xcm;
